scraper/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
psycopg2-binary==2.9.9

# HTTP client
httpx==0.27.2

# Browser automation (scraping)
playwright==1.47.0

# AI
anthropic==0.34.2

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
